Fix : fire ADD when new user connects
As mentionned in the description : Update the active peer list with a
new peer, given in parameter.

Or in the official javaDoc we can read that put returns : the previous
value associated with key, or null if there was no mapping for key.

The test has to be :
boolean update = mActivePeers.put(peer, new Date()) == null;

and not :
boolean update = mActivePeers.put(peer, new Date()) != null;

for a new peer.